+++ metabolights-referencelayer/dao/src/test/java/uk/ac/ebi/metabolights/referencelayer/DAO/db/SpeciesMembersDAOTest.java
@@ -75,7 +75,7 @@ public class SpeciesMembersDAOTest extends TestCase{
 		assertNotNull(spm);
 
 		// If the id is not null
-		if (expectedvalues[0] != null) 	assertEquals("Checking " + expectedvalues[1] + " id" , Long.parseLong(expectedvalues[0]) , spm.getId());
+		if (expectedvalues[0] != null) 	assertEquals("Checking " + expectedvalues[1] + " id" , Long.parseLong(expectedvalues[0]) , spm.getId().longValue());
 		assertEquals("Checking " + expectedvalues[1] + " Taxon" , expectedvalues[1] , spm.getTaxon());
 		assertEquals("Checking " + expectedvalues[1] + " Taxon description" , expectedvalues[2] , spm.getTaxonDesc());
